@@ -9,47 +9,51 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "copy-readme": "cp README.md dist/fedapay-checkout",
+    "serve:ssr:fedapay-Angular": "node dist/fedapay-angular/server/server.mjs",
     "build:lib": "ng-packagr -p ./projects/fedapay-checkout/ng-package.json && npm run copy-readme"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "core-js": "^2.5.4",
-    "node-sass": "^4.13.0",
-    "rxjs": "~6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "@angular/animations": "^18.2.0",
+    "@angular/common": "^18.2.0",
+    "@angular/compiler": "^18.2.0",
+    "@angular/core": "^18.2.0",
+    "@angular/forms": "^18.2.0",
+    "@angular/platform-browser": "^18.2.0",
+    "@angular/platform-browser-dynamic": "^18.2.0",
+    "@angular/platform-server": "^18.2.0",
+    "@angular/router": "^18.2.0",
+    "@angular/ssr": "^18.2.4",
+    "core-js": "^3.38.1",
+    "express": "^4.18.2",
+    "rxjs": "~7.8.0",
+    "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.0",
-    "@angular-devkit/build-ng-packagr": "~0.12.0",
-    "@angular/cli": "~7.2.4",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.2.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tsickle": ">=0.34.0",
-    "tslib": "^1.9.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "@angular-devkit/build-angular": "^18.2.4",
+    "@angular-devkit/build-ng-packagr": "^0.1002.0",
+    "@angular/cli": "^18.2.4",
+    "@angular/compiler-cli": "^18.2.0",
+    "@angular/language-service": "^18.2.4",
+    "@types/express": "^4.17.17",
+    "@types/jasmine": "~5.1.0",
+    "@types/jasminewd2": "^2.0.13",
+    "@types/node": "^18.18.0",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~5.2.0",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "ng-packagr": "^18.2.1",
+    "protractor": "^7.0.0",
+    "ts-node": "^10.9.2",
+    "tsickle": "^0.39.1",
+    "tslib": "^2.7.0",
+    "tslint": "^5.20.1",
+    "typescript": "~5.5.2"
   }
 }
